# apps/api/docs/swagger.yaml
definitions:
  github_com_projectx_api_internal_platform_httpx.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  internal_identity.Role:
    enum:
    - super_admin
    - studio_admin
    type: string
    x-enum-varnames:
    - RoleSuperAdmin
    - RoleStudioAdmin
  internal_identity.StudioBrand:
    properties:
      active:
        type: boolean
      brandColor:
        type: string
      logoUrl:
        type: string
      name:
        type: string
      slug:
        type: string
      socialPlannerEnabled:
        type: boolean
      subscriptionTier:
        type: string
    type: object
  internal_identity.changePasswordReq:
    properties:
      confirmPassword:
        type: string
      currentPassword:
        type: string
      newPassword:
        type: string
    type: object
  internal_identity.loginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_identity.meRes:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/internal_identity.Role'
      studio:
        $ref: '#/definitions/internal_identity.StudioBrand'
      studioId:
        type: string
    type: object
...
info:
  contact:
    name: Project-X support
  description: |-
    REST API for Project-X — the AI-run marketing + studio operations platform.
    Auth is a JWT stored in an HTTP-only cookie (set by POST /api/v1/auth/login).
    When calling protected endpoints from a tool like Swagger UI or curl, log in first and reuse the returned cookie.
  license:
    name: Proprietary
  title: Project-X API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Verifies credentials, issues a JWT, and sets it as an HttpOnly
        session cookie. Rate limited to 10 requests/minute per IP.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_identity.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_identity.meRes'
        "401":
          description: invalid email or password
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "422":
          description: email or password missing
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "429":
          description: too many attempts
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      summary: Log in with email and password
      tags:
      - Auth
...
  /api/v1/auth/me:
    get:
      description: Returns the authenticated user's profile and, if assigned to a
        studio, that studio's branding info. Clears the session cookie and returns
        401 if the studio backing the session no longer exists/is accessible.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_identity.meRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get current session user
      tags:
      - Auth
  /api/v1/auth/password:
    post:
      consumes:
      - application/json
      description: Validates the new password and confirmation, verifies the current
        password, then updates the stored password hash. Rate limited to 10 requests/minute
        per IP.
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_identity.changePasswordReq'
      responses:
        "204":
          description: no content
        "401":
          description: session invalid or current password incorrect for auth check
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "422":
          description: validation failed
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "429":
          description: too many attempts
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Change the current user's password
      tags:
      - Auth
...
  /api/v1/studios/{studioId}/users:
    get:
      description: Returns the id, email, and role of every user belonging to the
        given studio. Callers must be a super user or belong to the studio themselves.
      parameters:
      - description: Studio ID (UUID)
        in: path
        name: studioId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: missing or invalid studioId
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "403":
          description: cannot access this studio
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List users in a studio
      tags:
      - Auth
...
  /api/v1/webhooks/telegram/{botID}:
    post:
      consumes:
      - application/json
      description: Receives inbound Telegram bot updates for a specific bot. Looks
        up the channel by {botID}, then verifies the request's `X-Telegram-Bot-Api-Secret-Token`
        header against the per-channel secret registered via setWebhook before dispatching
        the update to the messaging service. Always responds 200 for unknown bots
        or malformed payloads so Telegram doesn't retry. No session/API-key auth —
        authenticated via the per-bot secret token header instead.
      parameters:
      - description: Telegram bot ID
        in: path
        name: botID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: missing bot id or bad request body
          schema:
            type: string
        "403":
          description: invalid secret token
          schema:
            type: string
      summary: Receive inbound Telegram bot webhook events
      tags:
      - Webhooks
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    description: JWT session cookie set by POST /api/v1/auth/login (cookie name configurable
      via COOKIE_NAME).
    in: cookie
    name: px_session
    type: apiKey
swagger: "2.0"